Work in progress
Profile Template is settled.

Still integrating Matthew's codes
